requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
pillow==10.0.0
PyQt5==5.15.9
mysql-connector-python==8.1.0
imutils==0.5.4
matplotlib==3.7.2
qrcode==7.4.2